requirements.txt
django~=3.2.5
djangorestframework~=3.12.4
djangorestframework-jwt~=1.11.0
psycopg2
psycopg2-binary==2.9.1
python-dotenv==0.19.0